requirements.txt
python-dotenv==1.0.1
discord.py==2.4.0
apscheduler==3.10.4
arxiv==2.1.3
pytz==2024.1
python-dateutil==2.9.0.post0
requests==2.32.3
pyyaml==6.0.2